Restructure module tree
