+++ src/tidytcells/mhc/__init__.py
@@ -1,0 +1,5 @@
+from ._main import (
+    standardise,
+    get_chain,
+    get_class
+)
