+++ src/tidytcells/tcr/_main.py
@@ -3,9 +3,7 @@ Utility functions related to TCRs and TCR genes.
 '''
 
 
-from .decomposed_gene import _DecomposedGene
-import json
-from pkg_resources import resource_stream
+from .._decomposed_gene import _DecomposedGene
 import re
 from typing import Optional
 from warnings import warn
@@ -14,19 +12,11 @@ from warnings import warn
 # --- STATIC RESOURCES ---
 
 
-with resource_stream(
-    __name__,
-    'resources/homosapiens_tcr.json') as s:
-    HOMOSAPIENS_TCR = json.load(s)
-with resource_stream(
-    __name__,
-    'resources/homosapiens_tcr_synonyms.json') as s:
-    HOMOSAPIENS_TCR_SYNONYMS = json.load(s)
-with resource_stream(
-    __name__,
-    'resources/musmusculus_tcr.json') as s:
-    MUSMUSCULUS_TCR = json.load(s)
-
+from .. import (
+    HOMOSAPIENS_TCR,
+    HOMOSAPIENS_TCR_SYNONYMS,
+    MUSMUSCULUS_TCR
+)
 PARSE_RE = re.compile(r'^([A-Z0-9\-\.\(\)\/]+)(\*([0-9]+))?')
